+++ arithmetization/src/test/java/net/consensys/linea/zktracer/forkSpecific/prague/floorprice/TrivialExecutionTests.java
@@ -24,6 +24,7 @@ import java.util.stream.Stream;
 import com.google.common.base.Preconditions;
 import net.consensys.linea.UnitTestWatcher;
 import net.consensys.linea.reporting.TracerTestBase;
+import net.consensys.linea.testing.AddressCollisions;
 import net.consensys.linea.testing.BytecodeCompiler;
 import net.consensys.linea.testing.BytecodeRunner;
 import net.consensys.linea.zktracer.module.txndata.cancun.transactions.CancunUserTransaction;
@@ -49,17 +50,17 @@ public class TrivialExecutionTests extends TracerTestBase {
       Bytes callData,
       boolean provideAccessList,
       DominantCost dominantCostPrediction,
-      BytecodeRunner.AddressCollisionCase addressCollisionCase,
+      AddressCollisions collision,
       TestInfo testInfo) {
     BytecodeRunner bytecodeRunner = BytecodeRunner.of(Bytes.EMPTY);
     AccessListEntry accessListEntry =
         new AccessListEntry(Address.fromHexString("0xABCD"), List.of());
     List<AccessListEntry> accessList = provideAccessList ? List.of(accessListEntry) : List.of();
-    if (addressCollisionCase == BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION) {
+    if (collision == AddressCollisions.NO_COLLISION) {
       bytecodeRunner.run(callData, accessList, chainConfig, testInfo);
     } else {
       bytecodeRunner.runWithAddressCollision(
-          callData, accessList, addressCollisionCase, chainConfig, testInfo);
+          callData, accessList, collision, chainConfig, testInfo);
     }
     // Test blocks contain 4 transactions: 2 system transactions, 1 user transaction (the one we
     // created) and 1 noop transaction.
@@ -77,7 +78,7 @@ public class TrivialExecutionTests extends TracerTestBase {
       Bytes callData,
       boolean provideAccessList,
       DominantCost dominantCostPrediction,
-      BytecodeRunner.AddressCollisionCase addressCollisionCase,
+      AddressCollisions collision,
       TestInfo testInfo) {
     BytecodeCompiler program = BytecodeCompiler.newProgram(chainConfig);
     program.op(OpCode.STOP);
@@ -85,11 +86,11 @@ public class TrivialExecutionTests extends TracerTestBase {
     AccessListEntry accessListEntry =
         new AccessListEntry(Address.fromHexString("0xABCD"), List.of());
     List<AccessListEntry> accessList = provideAccessList ? List.of(accessListEntry) : List.of();
-    if (addressCollisionCase == BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION) {
+    if (collision == AddressCollisions.NO_COLLISION) {
       bytecodeRunner.run(callData, accessList, chainConfig, testInfo);
     } else {
       bytecodeRunner.runWithAddressCollision(
-          callData, accessList, addressCollisionCase, chainConfig, testInfo);
+          callData, accessList, collision, chainConfig, testInfo);
     }
     // Test blocks contain 4 transactions: 2 system transactions, 1 user transaction (the one we
     // created) and 1 noop transaction.
@@ -116,13 +117,13 @@ public class TrivialExecutionTests extends TracerTestBase {
             buildCallData(CallDataSetting.ALL_ZEROS, true, 400),
             true,
             DominantCost.EXECUTION_COST_DOMINATES,
-            BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION));
+            AddressCollisions.NO_COLLISION));
     arguments.add(
         Arguments.of(
             buildCallData(CallDataSetting.ALL_ZEROS, true, 401),
             true,
             DominantCost.FLOOR_COST_DOMINATES,
-            BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION));
+            AddressCollisions.NO_COLLISION));
 
     // Case ALL_NON_ZEROS_EXCEPT_FOR_FIRST.
     // The transaction execution cost (TX_SKIP) is 21000 + 2400 + 16*length.
@@ -133,13 +134,13 @@ public class TrivialExecutionTests extends TracerTestBase {
             buildCallData(CallDataSetting.ALL_NON_ZEROS_EXCEPT_FOR_FIRST, false, 100),
             true,
             DominantCost.EXECUTION_COST_DOMINATES,
-            BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION));
+            AddressCollisions.NO_COLLISION));
     arguments.add(
         Arguments.of(
             buildCallData(CallDataSetting.ALL_NON_ZEROS_EXCEPT_FOR_FIRST, false, 101),
             true,
             DominantCost.FLOOR_COST_DOMINATES,
-            BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION));
+            AddressCollisions.NO_COLLISION));
 
     // Case ZEROS_AND_NON_ZEROS.
     // caveat for simplicity we consider even sizes.
@@ -151,13 +152,13 @@ public class TrivialExecutionTests extends TracerTestBase {
             buildCallData(CallDataSetting.ZEROS_AND_NON_ZEROS, false, 160),
             true,
             DominantCost.EXECUTION_COST_DOMINATES,
-            BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION));
+            AddressCollisions.NO_COLLISION));
     arguments.add(
         Arguments.of(
             buildCallData(CallDataSetting.ZEROS_AND_NON_ZEROS, false, 162),
             true,
             DominantCost.FLOOR_COST_DOMINATES,
-            BytecodeRunner.AddressCollisionCase.NO_IMPOSED_COLLISION));
+            AddressCollisions.NO_COLLISION));
 
     return arguments.stream();
   }
@@ -178,8 +179,7 @@ public class TrivialExecutionTests extends TracerTestBase {
     // The transaction execution cost (TX_SKIP) is 21000 + 4.
     // The floor cost is 21000 + 10.
 
-    for (BytecodeRunner.AddressCollisionCase collisionCase :
-        BytecodeRunner.AddressCollisionCase.values()) {
+    for (AddressCollisions collisionCase : AddressCollisions.values()) {
       arguments.add(
           Arguments.of(
               buildCallData(CallDataSetting.ALL_ZEROS, true, 1),
